Add delete confirm dialog

+++ src/app/pages/employee/employee.component.ts
@@ -1,18 +1,19 @@
 import { Component, OnInit, ViewChild, ViewEncapsulation } from '@angular/core';
 import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
-import { EmployeeAddComponent } from './dialog/employee-add/employee-add.component'; 
+import { EmployeeAddComponent } from './dialog/employee-add/employee-add.component';
 import { EmployeeEditComponent } from './dialog/employee-edit/employee-edit.component';
 import { Observable } from 'rxjs';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
 import { EmployeeService } from 'src/app/services/EmployeeService/employee.service';
+import { DeleteDialogService } from 'src/app/services/delete-dialog.service';
 
 @Component({
   selector: 'app-employee',
   templateUrl: './employee.component.html',
   styleUrls: ['./employee.component.scss']
 })
-export class EmployeeComponent implements OnInit{
+export class EmployeeComponent implements OnInit {
   @ViewChild(MatPaginator) paginator: MatPaginator;
 
   employeelist: any;
@@ -24,6 +25,7 @@ export class EmployeeComponent implements OnInit{
   constructor(
     private _dialog: MatDialog,
     private _employeeService: EmployeeService,
+    private _deleteDialogService: DeleteDialogService
   ) { }
 
 
@@ -36,7 +38,7 @@ export class EmployeeComponent implements OnInit{
     const filterValue = (event.target as HTMLInputElement).value;
     this.dataSource.filter = filterValue.trim().toLowerCase();
   }
-  
+
   GetEmployees() {
     this._employeeService.GetEmployeesList().subscribe(response => {
       this.employeelist = response.data;
@@ -46,13 +48,13 @@ export class EmployeeComponent implements OnInit{
     });
   }
 
-  OpenAddEmployee(){
+  OpenAddEmployee() {
     const dialogConfig = new MatDialogConfig()
     dialogConfig.width = "55%"
     const dialogRef = this._dialog.open(EmployeeAddComponent, dialogConfig);
     dialogRef.afterClosed().subscribe({
       next: (val) => {
-        if(val){
+        if (val) {
           this.GetEmployees();
         }
       }
@@ -74,13 +76,20 @@ export class EmployeeComponent implements OnInit{
   }
 
   DeleteEmployee(id: number) {
-    this._employeeService.DeleteEmployee(id).subscribe({
-      next: (res) => {
-        // this._coreService.openSnackBar('Employee Deleted!');
-        this.GetEmployees();
-      },
-      error: console.log,
-    })
+    this._deleteDialogService.openConfirmDialog("Do you really want to delete this record?")
+      .afterClosed().subscribe({
+        next: (val) => {
+          if (val) {
+            this._employeeService.DeleteEmployee(id).subscribe({
+              next: (res) => {
+                // this._coreService.openSnackBar('Employee Deleted!');
+                this.GetEmployees();
+              },
+              error: console.log,
+            })
+          }
+        }
+      });
   }
 }
